+++ packages/perseus/src/widgets/deprecated-standin/deprecated-standin.tsx
@@ -1,0 +1,43 @@
+import Banner from "@khanacademy/wonder-blocks-banner";
+import React from "react";
+
+import {PerseusI18nContext} from "../../components/i18n-context";
+import noopValidator from "../__shared__/noop-validator";
+
+import type {Widget, WidgetExports} from "../../types";
+
+// The props are type `any` on purpose so that this can receive props
+// from any deprecated widget
+type Props = any;
+
+// @ts-expect-error - TS2559 - Type 'DeprecatedStandin' has no properties in common with type 'Widget'.
+class DeprecatedStandin extends React.Component<Props> implements Widget {
+    static contextType = PerseusI18nContext;
+    declare context: React.ContextType<typeof PerseusI18nContext>;
+
+    render() {
+        return (
+            <div
+                style={{
+                    paddingTop: 8,
+                    paddingBottom: 8,
+                }}
+            >
+                <Banner
+                    text={this.context.strings.deprecatedStandin}
+                    kind="info"
+                    layout="full-width"
+                />
+            </div>
+        );
+    }
+}
+
+export default {
+    name: "deprecated-standin",
+    displayName: "Deprecated Standin",
+    widget: DeprecatedStandin,
+    hidden: true,
+    // TODO: things that aren't interactive shouldn't need validators
+    validator: () => noopValidator(1),
+} as WidgetExports<typeof DeprecatedStandin>;
